+++ gninasrc/lib/non_cache_cnn.cpp
@@ -55,11 +55,6 @@ fl non_cache_cnn::eval(model& m, fl v) const
 	return e;
 }
 
-non_cache_cnn::non_cache_cnn(const non_cache_cnn& src, CNNScorer& cnn_scorer_) : 
-    slope(src.slope), gd(src.gd), sgrid(src.sgrid), p(src.p), cnn_scorer(cnn_scorer_)
-{
-}
-
 fl non_cache_cnn::eval_deriv(model& m, fl v, const grid& user_grid) const
 {
 	fl e = 0;
