+++ gninasrc/lib/non_cache_cnn.h
@@ -28,13 +28,13 @@
 #include "non_cache.h"
 #include "cnn_scorer.h"
 
+struct parallel_mc_aux;
 struct non_cache_cnn : public non_cache {
 	non_cache_cnn(szv_grid_cache& gcache,
 		      const grid_dims& gd_,
 		      const precalculate* p_,
 		      fl slope_,
 		      CNNScorer& cnn_scorer_);
-    non_cache_cnn(const non_cache_cnn& src, CNNScorer& cnn_scorer_);
 	virtual ~non_cache_cnn() {}
 	virtual fl eval      (model& m, fl v) const; // needs m.coords
 	virtual fl eval_deriv(model& m, fl v, const grid& user_grid) const; // needs m.coords, sets m.minus_forces
@@ -43,6 +43,8 @@ struct non_cache_cnn : public non_cache {
 	fl getSlope() { return slope; }
 	virtual bool skip_interacting_pairs() const { return true; }
 	virtual bool adjust_center() const { return cnn_scorer.adjust_center(); }
+private:
+    friend struct parallel_mc_aux;
 protected:
 	fl slope;
 	grid_dims gd;
