+++ gninasrc/main/main.cpp
@@ -56,6 +56,7 @@
 #include <boost/lockfree/queue.hpp>
 #include <boost/unordered_map.hpp>
 #include "sem.h"
+#include "user_opts.h"
 
 #include <cuda_profiler_api.h>
